+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [master]
+### Added
+- Client.destroyMailboxes #47
 
 ## [0.0.15] - 2016-06-17
 ### Added